2 files modified, 1 added

+++ build.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
-rm -r dist
+if [[-d "dist"]]
+then
+  rm -r dist
+fi
+
 cd client
 npm run build
 mv dist ..
